Refactor crappy style and scale the image to a hardcoded size
This is in line with tutorial 05 on lazyfoo.net.  It turns out that
Rust's renderer.copy() - which is the equivalent of "blit" - defaults
to using the entire space rather than the original size of the texture.
So, rather than make a new rect to scale it up, we're making a new rect
to scale it down.

I'm going to try to document differences I notice between how things
work in C and how I think they should be translated into Rust.  Maybe
once I'm further in I can write a blog post about it or something.

Signed-off-by: Russell Currey <ruscur@russell.cc>