Lesson 08: geometry rendering
The biggest difference between C and Rust here is just that Point is now
a struct, and things that need points use the Point struct.  Unfortunately,
there's a bit of grossness here fiddling with whether things are unsigned
or signed.  It works out, though.  And look at all the freeing and whatnot
we don't need to care about in Rust!

Signed-off-by: Russell Currey <ruscur@russell.cc>